OCC: read edges/draft
git-svn-id: https://elsa-svn.onera.fr/CASSIOPEE/Trunk/Cassiopee/Apps/Modules@4543 2bbd7fe0-4929-0410-bb6b-e55522daae51

+++ OCC/OCC/CADviaOCC2.cpp
@@ -155,7 +155,7 @@ E_Int K_OCC::CADviaOCC::mesh_faces2
       
     // surface of revolution => duplicate, reverse and separate seams
     //bool is_of_revolution = ((E_Int)nodes.size() != connectB.cols());
-    bool is_of_revolution = (nodes.size() != connectB.cols());
+    bool is_of_revolution = (nodes.size() != (size_t)connectB.cols());
 #ifdef DEBUG_CAD_READER
     if (i == faulty_id && is_of_revolution) printf("Edge is not loop - Is of revolution=true\n");
 #endif
